SETproject/backend/pyproject.toml
[project]
name = "app-backend"
version = "0.1.0"
description = "ProjectSET backend application"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "sqlalchemy>=2.0.36",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "httpx>=0.27.0",
    "passlib[bcrypt]>=1.7.4",
    "bcrypt>=4.0,<5.0",
    "python-jose[cryptography]>=3.3.0",
    "python-multipart>=0.0.9",
    "email-validator>=2.2.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/app"]
